feat(analytics): implement sales dashboard API with SQL aggregation (#64)
Replace full-table-scan + Java stream pattern with targeted JPQL aggregate
queries (SUM/COUNT/AVG/GROUP BY) so analytics scale with data volume.

Key changes:
- OrderRepository: add sumRevenueBetween, countOrdersBetween,
  avgOrderValueBetween, findTopSellingProducts, sumRevenueGroupedByCategory,
  countDistinctCustomersByStatus, countReturningCustomers, sumRevenueByUser;
  remove broken calculateRevenueForPeriod (referenced 'COMPLETED' status
  which does not exist in the OrderStatus enum — always returned zero)
- UserRepository: add countNewUsersBetween
- SalesAnalyticsService: change Double monetary returns to BigDecimal
- SalesAnalyticsServiceImpl: implement top-selling products, revenue by
  category, and retention rate (previously placeholders); retention now
  computed from real data (returning customers / total customers)
- SalesDashboardDTO: fix averageOrderValue type from Double to BigDecimal
- SalesAnalyticsController: move to /api/v1/admin/analytics/sales
- Tests: replace mock-based unit tests with @DataJpaTest integration tests
  that exercise real JPQL queries against H2

Closes #64

+++ backend/src/main/java/com/example/buildnest_ecommerce/model/dto/SalesDashboardDTO.java
@@ -35,7 +35,7 @@ public class SalesDashboardDTO {
     private Long totalOrders;
 
     // Customer Metrics
-    private Double averageOrderValue;
+    private BigDecimal averageOrderValue;
     private Long totalCustomers;
     private Long newCustomersThisMonth;
     private Double customerRetentionRate;
